net: moved nested IP4::addr struct to independant struct ip4::Addr
