+++ api/net/ip4/ip4.hpp
@@ -1,6 +1,6 @@
 // This file is a part of the IncludeOS unikernel - www.includeos.org
 //
-// Copyright 2015 Oslo and Akershus University College of Applied Sciences
+// Copyright 2015-2016 Oslo and Akershus University College of Applied Sciences
 // and Alfred Bratterud
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,10 +18,9 @@
 #ifndef CLASS_IP4_HPP
 #define CLASS_IP4_HPP
 
-#include <regex>
-#include <string>
 #include <iostream>
 
+#include "addr.hpp"
 #include <common>
 #include <net/ethernet.hpp>
 #include <net/inet.hpp>
@@ -35,104 +34,15 @@ namespace net {
   /** IP4 layer */
   class IP4 {
   public:
+    using addr = ip4::Addr;
+
     /** Initialize. Sets a dummy linklayer out. */
     explicit IP4(Inet<LinkLayer, IP4>&) noexcept;
 
     /** Known transport layer protocols. */
     enum proto { IP4_ICMP=1, IP4_UDP=17, IP4_TCP=6 };
 
-    /** IP4 address representation */
-    struct addr {
-      uint32_t whole;
-      
-      addr() : whole(0) {} // uninitialized
-      addr(const uint32_t ipv4_addr)
-        : whole(ipv4_addr) {}
-      addr(const uint8_t p1, const uint8_t p2, const uint8_t p3, const uint8_t p4)
-        : whole(p1 | (p2 << 8) | (p3 << 16) | (p4 << 24)) {}
 
-      /**
-       * @brief Construct an IPv4 address from a {std::string}
-       * object
-       *
-       * @note If the {std::string} object doesn't contain a valid
-       * IPv4 representation then the instance will contain the
-       * address -> 0.0.0.0
-       *
-       * @param ipv4_addr:
-       * A {std::string} object representing an IPv4 address
-       */
-      addr(const std::string& ipv4_addr)
-        : addr{}
-      {
-      	Expects(ipv4_addr.size() >= 7 && ipv4_addr.size() <= 15); //< [7, 15] minimum and maximum address length
-
-        const static std::regex ipv4_address_pattern
-        {
-          "^(25[0–5]|2[0–4]\\d|[01]?\\d\\d?)\\."
-          "(25[0–5]|2[0–4]\\d|[01]?\\d\\d?)\\."
-          "(25[0–5]|2[0–4]\\d|[01]?\\d\\d?)\\."
-          "(25[0–5]|2[0–4]\\d|[01]?\\d\\d?)$"
-        };
-
-        std::smatch ipv4_parts;
-      
-        if (not std::regex_match(ipv4_addr, ipv4_parts, ipv4_address_pattern)) {
-          return;
-        }
-        
-        const auto p1 = static_cast<uint8_t>(std::stoi(ipv4_parts[1]));
-        const auto p2 = static_cast<uint8_t>(std::stoi(ipv4_parts[2]));
-        const auto p3 = static_cast<uint8_t>(std::stoi(ipv4_parts[3]));
-        const auto p4 = static_cast<uint8_t>(std::stoi(ipv4_parts[4]));
-
-        whole = p1 | (p2 << 8) | (p3 << 16) | (p4 << 24);
-      }
-      
-      inline addr& operator=(const addr cpy) noexcept {
-        whole = cpy.whole;
-        return *this;
-      }
-
-      /** Standard comparison operators */
-      bool operator==(const addr rhs)     const noexcept
-      { return whole == rhs.whole; }
-
-      bool operator==(const uint32_t rhs) const noexcept
-      { return  whole == rhs; }
-
-      bool operator<(const addr rhs)      const noexcept
-      { return whole < rhs.whole; }
-
-      bool operator<(const uint32_t rhs)  const noexcept
-      { return  whole < rhs; }
-
-      bool operator>(const addr rhs)      const noexcept
-      { return whole > rhs.whole; }
-
-      bool operator>(const uint32_t rhs)  const noexcept
-      { return  whole > rhs; }
-
-      bool operator!=(const addr rhs)     const noexcept
-      { return whole != rhs.whole; }
-
-      bool operator!=(const uint32_t rhs) const noexcept
-      { return  whole != rhs; }
-
-      addr operator & (const addr rhs)    const noexcept
-      { return addr(whole & rhs.whole); }
-      
-      /** x.x.x.x string representation */
-      std::string str() const {
-        char ipv4_addr[16];
-        sprintf(ipv4_addr, "%1i.%1i.%1i.%1i",
-                (whole >>  0) & 0xFF,
-                (whole >>  8) & 0xFF, 
-                (whole >> 16) & 0xFF, 
-                (whole >> 24) & 0xFF);
-        return ipv4_addr;
-      }
-    } __attribute__((packed)); //< IP4::addr
 
     static const addr INADDR_ANY;
     static const addr INADDR_BCAST;
